move derive_key/etc to util.py
