+++ src/wormhole/_encryption.py
@@ -1,58 +1,12 @@
-from hashlib import sha256
-
 from attr import attrib, attrs
 from attr.validators import instance_of
-from nacl import utils
-from nacl.exceptions import CryptoError
-from nacl.secret import SecretBox
 from spake2 import SPAKE2_Symmetric
 from zope.interface import implementer
 
 from . import _interfaces
 from .util import (bytes_to_dict, bytes_to_hexstr, dict_to_bytes,
-                   hexstr_to_bytes, to_bytes, HKDF, provides)
-
-CryptoError
-__all__ = ["derive_key", "derive_phase_key", "CryptoError", "Encryption"]
-
-
-def derive_key(key, purpose, length=SecretBox.KEY_SIZE):
-    if not isinstance(key, bytes):
-        raise TypeError(type(key))
-    if not isinstance(purpose, bytes):
-        raise TypeError(type(purpose))
-    if not isinstance(length, int):
-        raise TypeError(type(length))
-    return HKDF(key, length, CTXinfo=purpose)
-
-
-def derive_phase_key(key, side, phase):
-    assert isinstance(side, str), type(side)
-    assert isinstance(phase, str), type(phase)
-    side_bytes = side.encode("ascii")
-    phase_bytes = phase.encode("ascii")
-    purpose = (b"wormhole:phase:" + sha256(side_bytes).digest() +
-               sha256(phase_bytes).digest())
-    return derive_key(key, purpose)
-
-
-def decrypt_data(key, encrypted):
-    assert isinstance(key, bytes), type(key)
-    assert isinstance(encrypted, bytes), type(encrypted)
-    assert len(key) == SecretBox.KEY_SIZE, len(key)
-    box = SecretBox(key)
-    data = box.decrypt(encrypted)
-    return data
-
-
-def encrypt_data(key, plaintext):
-    assert isinstance(key, bytes), type(key)
-    assert isinstance(plaintext, bytes), type(plaintext)
-    assert len(key) == SecretBox.KEY_SIZE, len(key)
-    box = SecretBox(key)
-    nonce = utils.random(SecretBox.NONCE_SIZE)
-    return box.encrypt(plaintext, nonce)
-
+                   hexstr_to_bytes, to_bytes, provides, CryptoError,
+                   derive_key, derive_phase_key, decrypt_data, encrypt_data)
 
 @attrs
 @implementer(_interfaces.IEncryption)
